Split logic transform passes
