+++ src/transform/logic.rs
@@ -1,45 +1,22 @@
+mod buffers;
+mod compose;
+mod decompose;
+mod fold_or;
+mod optimize;
+
+#[cfg(test)]
+mod tests;
+
 use std::collections::{HashMap, HashSet, VecDeque};
 
 use disjoint_set::DisjointSet;
 use itertools::Itertools;
 
-use crate::graph::logic::{LogicGraph, LogicGraphBuilder};
-use crate::graph::{Graph, GraphNode, GraphNodeId, GraphNodeKind, SubGraphWithGraph};
-use crate::logic::{Logic, LogicType};
+use crate::graph::logic::LogicGraph;
+use crate::graph::{GraphNodeId, SubGraphWithGraph};
 
 pub struct LogicGraphTransformer {
     pub graph: LogicGraph,
-}
-
-#[derive(Clone, Default)]
-struct GraphPatternMatch {
-    bindings: HashMap<String, GraphNodeId>,
-    logic_nodes: HashMap<GraphNodeId, GraphNodeId>,
-}
-
-impl GraphPatternMatch {
-    fn bind_input(mut self, name: &str, graph_node_id: GraphNodeId) -> Option<Self> {
-        if let Some(existing) = self.bindings.get(name).copied() {
-            return (existing == graph_node_id).then_some(self);
-        }
-        self.bindings.insert(name.to_owned(), graph_node_id);
-        Some(self)
-    }
-
-    fn bind_logic(
-        mut self,
-        pattern_node_id: GraphNodeId,
-        graph_node_id: GraphNodeId,
-    ) -> Option<Self> {
-        if let Some(existing) = self.logic_nodes.get(&pattern_node_id).copied() {
-            return (existing == graph_node_id).then_some(self);
-        }
-        if self.logic_nodes.values().any(|id| *id == graph_node_id) {
-            return None;
-        }
-        self.logic_nodes.insert(pattern_node_id, graph_node_id);
-        Some(self)
-    }
 }
 
 impl LogicGraphTransformer {
@@ -49,1103 +26,6 @@ impl LogicGraphTransformer {
 
     pub fn finish(self) -> LogicGraph {
         self.graph
-    }
-
-    // |(a, b, c) => (a | b) | c
-    pub fn decompose_binops(&mut self) {
-        todo!()
-    }
-
-    // a & b => ~(a | b)
-    pub fn decompose_and(&mut self) -> eyre::Result<()> {
-        // check decomposable
-        if self.graph.graph.nodes.iter().any(|node| match &node.kind {
-            GraphNodeKind::Logic(logic) => match logic.logic_type {
-                LogicType::And if node.inputs.len() != 2 => false,
-                _ => false,
-            },
-            _ => false,
-        }) {
-            eyre::bail!("Cannot decompose! Run decompose binops before decomposing and!");
-        }
-
-        let target_nodes = self
-            .graph
-            .graph
-            .nodes
-            .iter()
-            .filter(|node| {
-                matches!(&node.kind,
-                    GraphNodeKind::Logic(logic) if matches!(logic.logic_type, LogicType::And),
-                )
-            })
-            .map(|node| node.id)
-            .collect_vec();
-
-        let and_gate = LogicGraphBuilder::new("~(~x|~y)".to_string())
-            .build("z".to_string())
-            .unwrap();
-
-        for node_id in target_nodes {
-            self.replace_binops_lazy(node_id, and_gate.graph.clone())?;
-        }
-
-        // for sort input sequence by id
-        self.graph.graph.build_inputs();
-        self.graph.graph.build_producers();
-        self.graph.graph.build_consumers();
-
-        Ok(())
-    }
-
-    // a ^ b => (~a & b) | (a & ~b)
-    pub fn decompose_xor(&mut self) -> eyre::Result<()> {
-        // check decomposable
-        if self.graph.graph.nodes.iter().any(|node| match &node.kind {
-            GraphNodeKind::Logic(logic) => match logic.logic_type {
-                LogicType::Xor if node.inputs.len() != 2 => false,
-                _ => false,
-            },
-            _ => false,
-        }) {
-            eyre::bail!("Cannot decompose! Run decompose binops before decomposing xor!");
-        }
-
-        let target_nodes = self
-            .graph
-            .graph
-            .nodes
-            .iter()
-            .filter(|node| {
-                matches!(&node.kind,
-                    GraphNodeKind::Logic(logic) if matches!(logic.logic_type, LogicType::Xor),
-                )
-            })
-            .map(|node| node.id)
-            .collect_vec();
-
-        let xor_gate = LogicGraphBuilder::new("(~x&y)|(x&~y)".to_string())
-            .build("z".to_string())
-            .unwrap();
-
-        for node_id in target_nodes {
-            self.replace_binops_lazy(node_id, xor_gate.graph.clone())?;
-        }
-
-        // for sort input sequence by id
-        self.graph.graph.build_inputs();
-        self.graph.graph.build_producers();
-        self.graph.graph.build_consumers();
-
-        Ok(())
-    }
-
-    // (a | b) | (c | d) => |(a, b, c, d)
-    pub fn fusion_orops(&mut self) {
-        todo!()
-    }
-
-    // (a & b) & (c & d) => &(a, b, c, d)
-    pub fn fusion_andops(&mut self) {
-        todo!()
-    }
-
-    // ~(~a) => a
-    pub fn remove_double_neg_expression(&mut self) {
-        let neg_nodes = self
-            .graph
-            .graph
-            .nodes
-            .iter()
-            .filter(|node| matches!(&node.kind, GraphNodeKind::Logic(Logic { logic_type }) if *logic_type == LogicType::Not))
-            .collect_vec();
-
-        let mut remove_target_ops = Vec::new();
-        let mut remove_targets = HashSet::new();
-
-        for node in neg_nodes {
-            if remove_targets.contains(&node.id) {
-                continue;
-            }
-
-            if node.outputs.is_empty() {
-                continue;
-            }
-
-            let Some(output) = self.graph.graph.find_node_by_id(node.outputs[0]) else {
-                continue;
-            };
-
-            if matches!(&output.kind, GraphNodeKind::Logic(Logic { logic_type }) if *logic_type == LogicType::Not)
-            {
-                remove_targets.insert(node.id);
-                remove_targets.insert(output.id);
-                remove_target_ops.push((node.id, output.id));
-            }
-        }
-
-        for (t1, t2) in remove_target_ops {
-            let Some(t1_node) = self.graph.graph.find_node_by_id(t1) else {
-                continue;
-            };
-            let Some(t2_node) = self.graph.graph.find_node_by_id(t2) else {
-                continue;
-            };
-
-            assert_eq!(t1_node.inputs.len(), 1);
-            let input = t1_node.inputs[0];
-
-            let outputs = t2_node.outputs.clone();
-
-            //   in                   in
-            //   |                    | \
-            // not(t1)                |  not(t1)
-            //   | \       =>         |    \
-            //   |  \                out1   out2
-            //   |   \
-            // not(t2)\
-            //   |     \
-            //  out1   out2
-            let reserve_t1 = t1_node.outputs.len() > 1;
-
-            if reserve_t1 {
-                self.graph
-                    .graph
-                    .find_node_by_id_mut(input)
-                    .unwrap()
-                    .outputs
-                    .extend(outputs.clone());
-            } else {
-                self.graph
-                    .graph
-                    .replace_target_output_node_ids(input, t1, outputs.clone());
-            }
-
-            for &output in &outputs {
-                self.graph
-                    .graph
-                    .replace_target_input_node_ids(output, t2, vec![input]);
-            }
-
-            if reserve_t1 {
-                self.graph
-                    .graph
-                    .find_node_by_id_mut(t1)
-                    .unwrap()
-                    .outputs
-                    .retain(|id| *id != t2);
-            } else {
-                self.graph.graph.remove_by_node_id_lazy(t1);
-            }
-            self.graph.graph.remove_by_node_id_lazy(t2);
-        }
-
-        self.graph.graph.build_producers();
-        self.graph.graph.build_consumers();
-    }
-
-    // replacable only (x, y) => (z)
-    fn replace_binops_lazy(&mut self, src: GraphNodeId, mut tar: Graph) -> eyre::Result<()> {
-        let g = &mut self.graph.graph;
-
-        let src_inputs = g.find_node_by_id(src).unwrap().inputs.clone();
-        let src_outputs = g.find_node_by_id(src).unwrap().outputs.clone();
-
-        let tar_inputs = tar.inputs();
-        let tar_inputs_outputs = tar
-            .inputs()
-            .iter()
-            .map(|node_id| tar.find_node_by_id(*node_id).unwrap().outputs.clone())
-            .collect_vec();
-        let tar_output = tar.outputs()[0];
-        let tar_output_input = tar.find_node_by_id(tar.outputs()[0]).unwrap().inputs[0];
-        let old_to_existing_ids = HashMap::from([
-            (tar_inputs[0], src_inputs[0]),
-            (tar_inputs[1], src_inputs[1]),
-        ]);
-
-        tar.remove_by_node_id_lazy(tar_output);
-        let old_to_current_ids = g.append_graph_with_replacements(tar, &old_to_existing_ids);
-
-        for index in 0..=1 {
-            if let Some(mut node) = g.find_node_by_id_mut(src_inputs[index]) {
-                node.outputs.extend(
-                    tar_inputs_outputs[index]
-                        .iter()
-                        .map(|id| old_to_current_ids.get(id).copied().unwrap_or(*id)),
-                );
-                node.outputs.retain(|node_id| *node_id != src);
-            }
-        }
-
-        let tar_output_input = old_to_current_ids
-            .get(&tar_output_input)
-            .copied()
-            .unwrap_or(tar_output_input);
-
-        g.replace_input_node_id_lazy(src, tar_output_input);
-        g.remove_by_node_id_lazy(src);
-        g.find_node_by_id_mut(tar_output_input).unwrap().outputs = src_outputs;
-
-        Ok(())
-    }
-
-    pub fn optimize(&mut self) -> eyre::Result<()> {
-        if self.graph.graph.outputs().len() != 1 {
-            eyre::bail!("You must split by outputs before run optimizing!");
-        }
-
-        // optimize logic graph using quine mccluskey
-        let input_nodes = self
-            .graph
-            .graph
-            .nodes
-            .iter()
-            .filter(|node| matches!(node.kind, GraphNodeKind::Input(_)))
-            .map(|node| (node.id, node.clone_node()))
-            .collect_vec();
-        let input_terms = input_nodes
-            .iter()
-            .enumerate()
-            .map(|(index, (node_id, _))| (*node_id, index as u8))
-            .collect::<HashMap<_, _>>();
-
-        fn make_qmc_form(
-            graph: &Graph,
-            input_terms: &HashMap<GraphNodeId, u8>,
-            node_id: GraphNodeId,
-        ) -> eyre::Result<quine_mc_cluskey::Bool> {
-            let node = graph.find_node_by_id(node_id).unwrap();
-            match &node.kind {
-                GraphNodeKind::Input(_) => Ok(quine_mc_cluskey::Bool::Term(input_terms[&node_id])),
-                GraphNodeKind::Output(_) => make_qmc_form(graph, input_terms, node.inputs[0]),
-                GraphNodeKind::Logic(logic) => Ok(match &logic.logic_type {
-                    LogicType::Not => quine_mc_cluskey::Bool::Not(Box::new(make_qmc_form(
-                        graph,
-                        input_terms,
-                        node.inputs[0],
-                    )?)),
-                    LogicType::And => quine_mc_cluskey::Bool::And(
-                        node.inputs
-                            .iter()
-                            .map(|input| make_qmc_form(graph, input_terms, *input).unwrap())
-                            .collect(),
-                    ),
-                    LogicType::Or => quine_mc_cluskey::Bool::Or(
-                        node.inputs
-                            .iter()
-                            .map(|input| make_qmc_form(graph, input_terms, *input).unwrap())
-                            .collect(),
-                    ),
-                    LogicType::Xor => unimplemented!(),
-                }),
-                _ => unreachable!(),
-            }
-        }
-
-        let results = make_qmc_form(
-            &self.graph.graph,
-            &input_terms,
-            self.graph.graph.outputs()[0],
-        )?
-        .simplify();
-
-        fn make_rc_form(
-            graph: &mut Graph,
-            lookup: &Vec<(GraphNodeId, GraphNode)>,
-            node: &quine_mc_cluskey::Bool,
-        ) -> GraphNodeId {
-            let node = match node {
-                quine_mc_cluskey::Bool::Term(v) => GraphNode {
-                    kind: GraphNodeKind::Input(lookup[*v as usize].1.kind.as_input().to_owned()),
-                    ..Default::default()
-                },
-                quine_mc_cluskey::Bool::And(op) => GraphNode {
-                    kind: GraphNodeKind::Logic(Logic {
-                        logic_type: LogicType::And,
-                    }),
-                    inputs: op
-                        .iter()
-                        .map(|v| make_rc_form(graph, lookup, v))
-                        .collect_vec(),
-                    ..Default::default()
-                },
-                quine_mc_cluskey::Bool::Or(op) => GraphNode {
-                    kind: GraphNodeKind::Logic(Logic {
-                        logic_type: LogicType::Or,
-                    }),
-                    inputs: op
-                        .iter()
-                        .map(|v| make_rc_form(graph, lookup, v))
-                        .collect_vec(),
-                    ..Default::default()
-                },
-                quine_mc_cluskey::Bool::Not(v) => GraphNode {
-                    kind: GraphNodeKind::Logic(Logic {
-                        logic_type: LogicType::Not,
-                    }),
-                    inputs: vec![make_rc_form(graph, lookup, v)],
-                    ..Default::default()
-                },
-                _ => unreachable!(),
-            };
-
-            if let GraphNodeKind::Input(name) = &node.kind {
-                if let Some(existing) = graph.nodes.iter().find(
-                    |existing| matches!(&existing.kind, GraphNodeKind::Input(other) if name == other),
-                ) {
-                    return existing.id;
-                }
-            }
-
-            graph.add_node(node)
-        }
-
-        let mut graph = Graph::default();
-        let node = make_rc_form(&mut graph, &input_nodes, &results[0]);
-        graph.add_node(GraphNode {
-            kind: GraphNodeKind::Output(
-                self.graph
-                    .graph
-                    .find_node_by_id(self.graph.graph.outputs()[0])
-                    .unwrap()
-                    .kind
-                    .as_output()
-                    .to_owned(),
-            ),
-            inputs: vec![node],
-            ..Default::default()
-        });
-        self.graph.graph = graph;
-        self.graph.graph.build_outputs();
-        self.graph.graph.build_producers();
-        self.graph.graph.build_consumers();
-
-        Ok(())
-    }
-
-    // Merge structurally equivalent logic nodes so shared sub-expressions are
-    // placed once and reused by all consumers. This is intentionally structural
-    // CSE only: it canonicalizes commutative input order, but does not apply
-    // boolean algebra rewrites such as idempotence or absorption.
-    pub fn optimize_cse(&mut self) -> eyre::Result<()> {
-        let mut expressions: HashMap<GraphNodeId, String> = HashMap::new();
-        let mut representatives: HashMap<String, GraphNodeId> = HashMap::new();
-        let mut replacements = Vec::new();
-
-        for node_id in self.graph.graph.topological_order() {
-            let Some(node) = self.graph.graph.find_node_by_id(node_id) else {
-                continue;
-            };
-
-            let expression = match &node.kind {
-                GraphNodeKind::Input(name) => format!("Input({name})"),
-                GraphNodeKind::Logic(logic) => {
-                    let mut inputs = node
-                        .inputs
-                        .iter()
-                        .map(|input| {
-                            expressions.get(input).cloned().ok_or_else(|| {
-                                eyre::eyre!("CSE expression is missing for input node {input}")
-                            })
-                        })
-                        .collect::<eyre::Result<Vec<_>>>()?;
-
-                    if matches!(
-                        logic.logic_type,
-                        LogicType::Or | LogicType::And | LogicType::Xor
-                    ) {
-                        inputs.sort();
-                    }
-
-                    format!("{}({})", logic.logic_type.name(), inputs.join(", "))
-                }
-                GraphNodeKind::Sequential(sequential) => {
-                    format!("Sequential#{}({})", node.id, sequential.name())
-                }
-                GraphNodeKind::Output(_) => continue,
-                _ => continue,
-            };
-
-            expressions.insert(node_id, expression.clone());
-
-            if !node.kind.is_logic() {
-                continue;
-            }
-
-            if let Some(&representative) = representatives.get(&expression) {
-                replacements.push((node_id, representative));
-            } else {
-                representatives.insert(expression, node_id);
-            }
-        }
-
-        for (from, to) in replacements {
-            self.graph.graph.replace_input_node_id_lazy(from, to);
-            self.graph.graph.remove_by_node_id_lazy(from);
-        }
-
-        self.graph.graph.build_outputs();
-        self.graph.graph.build_inputs();
-        self.graph.graph.build_outputs();
-        self.graph.graph.build_producers();
-        self.graph.graph.build_consumers();
-        self.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    pub fn compose_high_level_gates(&mut self) -> eyre::Result<()> {
-        self.compose_gate_pattern(("(~(x|~y))|(~(y|~x))", LogicType::Xor, "composed-xor"))?;
-        self.compose_gate_pattern(("~(~x|~y)", LogicType::And, "composed-and"))?;
-        self.optimize_cse()?;
-
-        for pattern in [
-            ("(~x&y)|(x&~y)", LogicType::Xor, "composed-xor"),
-            ("(x&(~x|~y))|(y&(~x|~y))", LogicType::Xor, "composed-xor"),
-            ("(x&(~x|~y))|~((x&y)|~y)", LogicType::Xor, "composed-xor"),
-        ] {
-            self.compose_gate_pattern(pattern)?;
-        }
-        self.compose_truth_table_gates()?;
-        self.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    fn compose_gate_pattern(
-        &mut self,
-        (pattern_stmt, logic_type, tag): (&str, LogicType, &str),
-    ) -> eyre::Result<()> {
-        let pattern_graph = Self::optimized_pattern_graph(pattern_stmt)?;
-        let pattern_inputs = pattern_graph
-            .inputs()
-            .into_iter()
-            .map(|id| {
-                pattern_graph
-                    .find_node_by_id(id)
-                    .unwrap()
-                    .kind
-                    .as_input()
-                    .clone()
-            })
-            .collect_vec();
-        let pattern_root_id = pattern_graph
-            .find_node_by_id(pattern_graph.outputs()[0])
-            .unwrap()
-            .inputs[0];
-        let root_logic_type = pattern_graph
-            .find_node_by_id(pattern_root_id)
-            .unwrap()
-            .kind
-            .as_logic()
-            .unwrap()
-            .logic_type;
-        let candidate_ids = self
-            .graph
-            .graph
-            .nodes
-            .iter()
-            .filter(|node| {
-                matches!(&node.kind, GraphNodeKind::Logic(logic) if logic.logic_type == root_logic_type)
-            })
-            .map(|node| node.id)
-            .collect_vec();
-
-        for node_id in candidate_ids {
-            let Some(pattern_match) = self.match_graph_pattern(
-                &pattern_graph,
-                pattern_root_id,
-                node_id,
-                GraphPatternMatch::default(),
-            ) else {
-                continue;
-            };
-
-            let Some(inputs) = pattern_inputs
-                .iter()
-                .map(|name| pattern_match.bindings.get(name).copied())
-                .collect::<Option<Vec<_>>>()
-            else {
-                continue;
-            };
-            if !inputs.iter().all_unique() {
-                continue;
-            }
-
-            let Some(outputs) = self
-                .graph
-                .graph
-                .find_node_by_id(node_id)
-                .map(|node| node.outputs.clone())
-            else {
-                continue;
-            };
-
-            let removals = self.removable_pattern_nodes(&pattern_match.logic_nodes, node_id);
-            self.graph.graph.replace_nodes_with(
-                &removals,
-                GraphNodeKind::Logic(Logic { logic_type }),
-                inputs,
-                outputs,
-                tag.to_owned(),
-            );
-        }
-
-        Ok(())
-    }
-
-    fn optimized_pattern_graph(pattern_stmt: &str) -> eyre::Result<Graph> {
-        let pattern_graph =
-            LogicGraphBuilder::new(pattern_stmt.to_owned()).build("z".to_owned())?;
-        let mut transformer = LogicGraphTransformer::new(pattern_graph);
-        transformer.optimize_cse()?;
-        Ok(transformer.finish().graph)
-    }
-
-    fn compose_truth_table_gates(&mut self) -> eyre::Result<()> {
-        let candidate_ids = self
-            .graph
-            .graph
-            .nodes
-            .iter()
-            .filter(|node| {
-                matches!(&node.kind, GraphNodeKind::Logic(_)) && !node.tag.starts_with("composed-")
-            })
-            .map(|node| node.id)
-            .collect_vec();
-
-        for node_id in candidate_ids {
-            let Some((logic_type, inputs, removals)) = self.truth_table_gate_candidate(node_id)?
-            else {
-                continue;
-            };
-            let Some(outputs) = self
-                .graph
-                .graph
-                .find_node_by_id(node_id)
-                .map(|node| node.outputs.clone())
-            else {
-                continue;
-            };
-
-            let tag = Self::composed_gate_tag(logic_type);
-            let inputs = self.binary_gate_inputs_for_replacement(logic_type, &inputs, tag);
-            self.graph.graph.replace_nodes_with(
-                &removals,
-                GraphNodeKind::Logic(Logic { logic_type }),
-                inputs,
-                outputs,
-                tag.to_owned(),
-            );
-        }
-
-        Ok(())
-    }
-
-    fn truth_table_gate_candidate(
-        &self,
-        root_id: GraphNodeId,
-    ) -> eyre::Result<Option<(LogicType, Vec<GraphNodeId>, HashSet<GraphNodeId>)>> {
-        if self.graph.graph.find_node_by_id(root_id).is_none() {
-            return Ok(None);
-        }
-
-        let mut cone_nodes = HashSet::new();
-        let mut inputs = HashSet::new();
-        self.collect_truth_table_cone(root_id, &mut cone_nodes, &mut inputs)?;
-
-        let mut inputs = inputs.into_iter().collect_vec();
-        inputs.sort_unstable();
-        if !(2..=3).contains(&inputs.len()) {
-            return Ok(None);
-        }
-
-        let table = self.evaluate_cone_truth_table(root_id, &inputs)?;
-        let logic_type = if Self::is_xor_truth_table(&table, inputs.len()) {
-            LogicType::Xor
-        } else if inputs.len() == 2 && Self::is_and_truth_table(&table, inputs.len()) {
-            LogicType::And
-        } else {
-            return Ok(None);
-        };
-
-        Ok(Some((
-            logic_type,
-            inputs,
-            self.removable_cone_nodes(&cone_nodes, root_id),
-        )))
-    }
-
-    fn collect_truth_table_cone(
-        &self,
-        node_id: GraphNodeId,
-        cone_nodes: &mut HashSet<GraphNodeId>,
-        inputs: &mut HashSet<GraphNodeId>,
-    ) -> eyre::Result<()> {
-        let Some(node) = self.graph.graph.find_node_by_id(node_id) else {
-            eyre::bail!("truth-table compose references missing node {node_id}");
-        };
-
-        if !matches!(&node.kind, GraphNodeKind::Logic(_)) {
-            inputs.insert(node_id);
-            return Ok(());
-        }
-
-        if !cone_nodes.insert(node_id) {
-            return Ok(());
-        }
-
-        for input in &node.inputs {
-            self.collect_truth_table_cone(*input, cone_nodes, inputs)?;
-        }
-
-        Ok(())
-    }
-
-    fn evaluate_cone_truth_table(
-        &self,
-        root_id: GraphNodeId,
-        inputs: &[GraphNodeId],
-    ) -> eyre::Result<Vec<bool>> {
-        (0..(1usize << inputs.len()))
-            .map(|mask| {
-                let input_values = inputs
-                    .iter()
-                    .enumerate()
-                    .map(|(index, input)| (*input, (mask & (1 << index)) != 0))
-                    .collect::<HashMap<_, _>>();
-                self.evaluate_cone_node(root_id, &input_values, &mut HashMap::new())
-            })
-            .collect()
-    }
-
-    fn evaluate_cone_node(
-        &self,
-        node_id: GraphNodeId,
-        input_values: &HashMap<GraphNodeId, bool>,
-        memo: &mut HashMap<GraphNodeId, bool>,
-    ) -> eyre::Result<bool> {
-        if let Some(value) = input_values.get(&node_id) {
-            return Ok(*value);
-        }
-        if let Some(value) = memo.get(&node_id) {
-            return Ok(*value);
-        }
-
-        let Some(node) = self.graph.graph.find_node_by_id(node_id) else {
-            eyre::bail!("truth-table compose evaluates missing node {node_id}");
-        };
-        let Some(logic) = node.kind.as_logic() else {
-            eyre::bail!("truth-table compose reached unbound leaf node {node_id}");
-        };
-        let values = node
-            .inputs
-            .iter()
-            .map(|input| self.evaluate_cone_node(*input, input_values, memo))
-            .collect::<eyre::Result<Vec<_>>>()?;
-        let value = match logic.logic_type {
-            LogicType::Not => !values[0],
-            LogicType::And => values.iter().all(|value| *value),
-            LogicType::Or => values.iter().any(|value| *value),
-            LogicType::Xor => values.iter().filter(|value| **value).count() % 2 == 1,
-        };
-
-        memo.insert(node_id, value);
-        Ok(value)
-    }
-
-    fn is_xor_truth_table(table: &[bool], input_count: usize) -> bool {
-        table
-            .iter()
-            .enumerate()
-            .all(|(mask, value)| *value == (mask.count_ones() as usize % 2 == 1))
-            && input_count >= 2
-    }
-
-    fn is_and_truth_table(table: &[bool], input_count: usize) -> bool {
-        let all_enabled = (1usize << input_count) - 1;
-        table
-            .iter()
-            .enumerate()
-            .all(|(mask, value)| *value == (mask == all_enabled))
-    }
-
-    fn composed_gate_tag(logic_type: LogicType) -> &'static str {
-        match logic_type {
-            LogicType::And => "composed-and",
-            LogicType::Xor => "composed-xor",
-            _ => unreachable!(),
-        }
-    }
-
-    fn binary_gate_inputs_for_replacement(
-        &mut self,
-        logic_type: LogicType,
-        inputs: &[GraphNodeId],
-        tag: &str,
-    ) -> Vec<GraphNodeId> {
-        if inputs.len() <= 2 {
-            return inputs.to_vec();
-        }
-
-        let mut current =
-            self.add_connected_logic_node(logic_type, vec![inputs[0], inputs[1]], tag);
-        for input in inputs.iter().take(inputs.len() - 1).skip(2) {
-            current = self.add_connected_logic_node(logic_type, vec![current, *input], tag);
-        }
-
-        vec![current, *inputs.last().unwrap()]
-    }
-
-    fn add_connected_logic_node(
-        &mut self,
-        logic_type: LogicType,
-        inputs: Vec<GraphNodeId>,
-        tag: &str,
-    ) -> GraphNodeId {
-        let node_id = self.graph.graph.add_node(GraphNode {
-            kind: GraphNodeKind::Logic(Logic { logic_type }),
-            inputs: inputs.clone(),
-            tag: tag.to_owned(),
-            ..Default::default()
-        });
-
-        for input in inputs {
-            if let Some(mut input_node) = self.graph.graph.find_node_by_id_mut(input) {
-                if !input_node.outputs.contains(&node_id) {
-                    input_node.outputs.push(node_id);
-                }
-            }
-        }
-
-        node_id
-    }
-
-    fn match_graph_pattern(
-        &self,
-        pattern_graph: &Graph,
-        pattern_node_id: GraphNodeId,
-        graph_node_id: GraphNodeId,
-        pattern_match: GraphPatternMatch,
-    ) -> Option<GraphPatternMatch> {
-        let pattern_node = pattern_graph.find_node_by_id(pattern_node_id)?;
-        if let GraphNodeKind::Input(name) = &pattern_node.kind {
-            return pattern_match.bind_input(name, graph_node_id);
-        }
-
-        let pattern_logic = pattern_node.kind.as_logic()?;
-        let graph_node = self.graph.graph.find_node_by_id(graph_node_id)?;
-        let graph_logic = graph_node.kind.as_logic()?;
-        let pattern_inputs = pattern_node.inputs.clone();
-        let graph_inputs = graph_node.inputs.clone();
-
-        if pattern_logic.logic_type != graph_logic.logic_type
-            || pattern_inputs.len() != graph_inputs.len()
-        {
-            return None;
-        }
-
-        let pattern_match = pattern_match.bind_logic(pattern_node_id, graph_node_id)?;
-        Self::ordered_logic_inputs(pattern_logic.logic_type, graph_inputs)
-            .into_iter()
-            .find_map(|graph_inputs| {
-                self.match_pattern_inputs(
-                    pattern_graph,
-                    &pattern_inputs,
-                    graph_inputs,
-                    pattern_match.clone(),
-                )
-            })
-    }
-
-    fn ordered_logic_inputs(
-        logic_type: LogicType,
-        graph_inputs: Vec<GraphNodeId>,
-    ) -> Vec<Vec<GraphNodeId>> {
-        if matches!(logic_type, LogicType::Or | LogicType::And | LogicType::Xor)
-            && graph_inputs.len() == 2
-            && graph_inputs[0] != graph_inputs[1]
-        {
-            let reversed_inputs = vec![graph_inputs[1], graph_inputs[0]];
-            return vec![graph_inputs, reversed_inputs];
-        }
-        vec![graph_inputs]
-    }
-
-    fn match_pattern_inputs(
-        &self,
-        pattern_graph: &Graph,
-        pattern_inputs: &[GraphNodeId],
-        graph_inputs: Vec<GraphNodeId>,
-        mut pattern_match: GraphPatternMatch,
-    ) -> Option<GraphPatternMatch> {
-        for (pattern_input, graph_input) in pattern_inputs.iter().zip(graph_inputs) {
-            pattern_match = self.match_graph_pattern(
-                pattern_graph,
-                *pattern_input,
-                graph_input,
-                pattern_match,
-            )?;
-        }
-        Some(pattern_match)
-    }
-
-    fn removable_pattern_nodes(
-        &self,
-        matched_logic_nodes: &HashMap<GraphNodeId, GraphNodeId>,
-        root_id: GraphNodeId,
-    ) -> HashSet<GraphNodeId> {
-        let matched_ids = matched_logic_nodes.values().copied().collect_vec();
-        let mut removals = HashSet::from([root_id]);
-
-        loop {
-            let mut changed = false;
-            for node_id in &matched_ids {
-                if removals.contains(node_id) {
-                    continue;
-                }
-                let Some(node) = self.graph.graph.find_node_by_id(*node_id) else {
-                    continue;
-                };
-                if node.outputs.iter().all(|output| removals.contains(output)) {
-                    removals.insert(*node_id);
-                    changed = true;
-                }
-            }
-            if !changed {
-                break;
-            }
-        }
-
-        removals
-    }
-
-    fn removable_cone_nodes(
-        &self,
-        cone_nodes: &HashSet<GraphNodeId>,
-        root_id: GraphNodeId,
-    ) -> HashSet<GraphNodeId> {
-        let mut removals = HashSet::from([root_id]);
-
-        loop {
-            let mut changed = false;
-            for node_id in cone_nodes {
-                if removals.contains(node_id) {
-                    continue;
-                }
-                let Some(node) = self.graph.graph.find_node_by_id(*node_id) else {
-                    continue;
-                };
-                if node.outputs.iter().all(|output| removals.contains(output)) {
-                    removals.insert(*node_id);
-                    changed = true;
-                }
-            }
-            if !changed {
-                break;
-            }
-        }
-
-        removals
-    }
-
-    pub fn fold_or_chains(&mut self) -> eyre::Result<()> {
-        let or_node_ids = self
-            .graph
-            .graph
-            .nodes
-            .iter()
-            .filter(|node| {
-                matches!(&node.kind, GraphNodeKind::Logic(logic) if logic.logic_type == LogicType::Or)
-            })
-            .map(|node| node.id)
-            .collect_vec();
-
-        for node_id in or_node_ids {
-            let Some(node) = self.graph.graph.find_node_by_id(node_id) else {
-                continue;
-            };
-            if !matches!(&node.kind, GraphNodeKind::Logic(logic) if logic.logic_type == LogicType::Or)
-            {
-                continue;
-            }
-
-            let mut folded_inputs = Vec::new();
-            let mut removable_or_nodes = HashSet::new();
-
-            for input in node.inputs.clone() {
-                Self::collect_folded_or_inputs(
-                    &self.graph.graph,
-                    input,
-                    node_id,
-                    &mut folded_inputs,
-                    &mut removable_or_nodes,
-                    &mut HashSet::new(),
-                )?;
-            }
-
-            folded_inputs.sort_unstable();
-            folded_inputs.dedup();
-
-            match folded_inputs.len() {
-                0 => {}
-                1 => {
-                    let input = folded_inputs[0];
-                    self.graph.graph.replace_node_id_lazy(node_id, input);
-                    removable_or_nodes.insert(node_id);
-                }
-                _ => {
-                    let inputs = self.binary_or_inputs_ending_at(&folded_inputs);
-                    if let Some(mut node) = self.graph.graph.find_node_by_id_mut(node_id) {
-                        node.inputs = inputs;
-                    }
-                }
-            }
-
-            for removable in removable_or_nodes {
-                self.graph.graph.remove_by_node_id_lazy(removable);
-            }
-
-            self.graph.graph.build_outputs();
-            self.graph.graph.build_inputs();
-            self.graph.graph.build_outputs();
-            self.graph.graph.build_producers();
-            self.graph.graph.build_consumers();
-        }
-
-        self.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    fn binary_or_inputs_ending_at(&mut self, inputs: &[GraphNodeId]) -> Vec<GraphNodeId> {
-        assert!(inputs.len() >= 2);
-        if inputs.len() == 2 {
-            return inputs.to_vec();
-        }
-
-        let mut current = self.graph.graph.add_node(GraphNode {
-            kind: GraphNodeKind::Logic(Logic {
-                logic_type: LogicType::Or,
-            }),
-            inputs: vec![inputs[0], inputs[1]],
-            tag: "folded-or".to_owned(),
-            ..Default::default()
-        });
-
-        for input in inputs.iter().take(inputs.len() - 1).skip(2) {
-            current = self.graph.graph.add_node(GraphNode {
-                kind: GraphNodeKind::Logic(Logic {
-                    logic_type: LogicType::Or,
-                }),
-                inputs: vec![current, *input],
-                tag: "folded-or".to_owned(),
-                ..Default::default()
-            });
-        }
-
-        vec![current, *inputs.last().unwrap()]
-    }
-
-    fn collect_folded_or_inputs(
-        graph: &Graph,
-        input_id: GraphNodeId,
-        consumer_id: GraphNodeId,
-        folded_inputs: &mut Vec<GraphNodeId>,
-        removals: &mut HashSet<GraphNodeId>,
-        path: &mut HashSet<GraphNodeId>,
-    ) -> eyre::Result<()> {
-        if !path.insert(input_id) {
-            eyre::bail!("cannot fold OR chain through cyclic node {input_id}");
-        }
-
-        let Some(node) = graph.find_node_by_id(input_id) else {
-            eyre::bail!("cannot fold OR chain through missing node {input_id}");
-        };
-
-        if !matches!(&node.kind, GraphNodeKind::Logic(logic) if logic.logic_type == LogicType::Or) {
-            folded_inputs.push(input_id);
-            path.remove(&input_id);
-            return Ok(());
-        }
-
-        let can_remove_current = node.outputs.iter().all(|output| *output == consumer_id);
-        if can_remove_current {
-            removals.insert(input_id);
-        }
-
-        for nested_input in &node.inputs {
-            Self::collect_folded_or_inputs(
-                graph,
-                *nested_input,
-                input_id,
-                folded_inputs,
-                removals,
-                path,
-            )?;
-        }
-
-        path.remove(&input_id);
-        Ok(())
-    }
-
-    pub fn insert_buffers_for_direct_or_to_or(&mut self) -> eyre::Result<()> {
-        let mut direct_edges = Vec::new();
-        for node in self.graph.graph.nodes.iter() {
-            if !matches!(&node.kind, GraphNodeKind::Logic(logic) if logic.logic_type == LogicType::Or)
-            {
-                continue;
-            }
-            for &output_id in &node.outputs {
-                let Some(output) = self.graph.graph.find_node_by_id(output_id) else {
-                    continue;
-                };
-                if matches!(&output.kind, GraphNodeKind::Logic(logic) if logic.logic_type == LogicType::Or)
-                {
-                    direct_edges.push((node.id, output.id));
-                }
-            }
-        }
-
-        for (from, to) in direct_edges {
-            let first_not = self.graph.graph.add_node(GraphNode {
-                kind: GraphNodeKind::Logic(Logic {
-                    logic_type: LogicType::Not,
-                }),
-                inputs: vec![from],
-                tag: "auto-buffer".to_owned(),
-                ..Default::default()
-            });
-            let second_not = self.graph.graph.add_node(GraphNode {
-                kind: GraphNodeKind::Logic(Logic {
-                    logic_type: LogicType::Not,
-                }),
-                inputs: vec![first_not],
-                outputs: vec![to],
-                tag: "auto-buffer".to_owned(),
-            });
-            self.graph
-                .graph
-                .find_node_by_id_mut(first_not)
-                .unwrap()
-                .outputs = vec![second_not];
-
-            self.graph
-                .graph
-                .replace_target_output_node_ids(from, to, vec![first_not]);
-            self.graph
-                .graph
-                .replace_target_input_node_ids(to, from, vec![second_not]);
-        }
-
-        self.graph.graph.build_inputs();
-        self.graph.graph.build_outputs();
-        self.graph.graph.build_producers();
-        self.graph.graph.build_consumers();
-        self.graph.graph.verify()?;
-
-        Ok(())
     }
 
     pub fn cluster(&self, include_ouput_node: bool) -> Vec<SubGraphWithGraph> {
@@ -1214,328 +94,5 @@ impl LogicGraphTransformer {
             .into_values()
             .map(|nodes| SubGraphWithGraph::from(&self.graph.graph, nodes))
             .collect_vec()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::graph::analysis::equivalent_expression_groups;
-    use crate::graph::logic::predefined_logics;
-    use crate::graph::world::WorldGraphBuilder;
-    use crate::nbt::NBTRoot;
-    use crate::sequential::{SequentialPrimitive, SequentialType};
-    use crate::transform::world_to_logic::WorldToLogicTransformer;
-
-    fn has_duplicate_expression(graph: &LogicGraph, expression: &str) -> bool {
-        equivalent_expression_groups(graph)
-            .iter()
-            .any(|group| group.expression == expression)
-    }
-
-    #[test]
-    fn optimize_cse_merges_duplicate_not_nodes() -> eyre::Result<()> {
-        let graph = LogicGraph::from_stmt("~a|~a", "out")?;
-        let mut transform = LogicGraphTransformer::new(graph);
-
-        transform.optimize_cse()?;
-
-        assert!(!has_duplicate_expression(&transform.graph, "Not(Input(a))"));
-        transform.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn optimize_cse_merges_commutative_or_nodes() -> eyre::Result<()> {
-        let graph = LogicGraph::from_stmt("(a|b)|(b|a)", "out")?;
-        let mut transform = LogicGraphTransformer::new(graph);
-
-        transform.optimize_cse()?;
-
-        assert!(!has_duplicate_expression(
-            &transform.graph,
-            "Or(Input(a), Input(b))"
-        ));
-        transform.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn optimize_cse_rewrites_outputs_to_representative() -> eyre::Result<()> {
-        let out_x = LogicGraph::from_stmt("~a", "x")?;
-        let out_y = LogicGraph::from_stmt("~a", "y")?;
-
-        let mut graph = out_x.clone();
-        graph.graph.merge(out_y.graph);
-
-        let mut transform = LogicGraphTransformer::new(graph);
-        transform.optimize_cse()?;
-
-        let output_inputs = transform
-            .graph
-            .graph
-            .nodes
-            .iter()
-            .filter_map(|node| match &node.kind {
-                GraphNodeKind::Output(name) if name == "x" || name == "y" => Some(node.inputs[0]),
-                _ => None,
-            })
-            .collect_vec();
-
-        assert_eq!(output_inputs.len(), 2);
-        assert_eq!(output_inputs[0], output_inputs[1]);
-        transform.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn fold_or_chains_preserves_binary_or_fan_in() -> eyre::Result<()> {
-        let mut graph = LogicGraph::from_stmt("a|b", "x")?;
-        graph.graph.merge_by_outin(
-            LogicGraph::from_stmt("x|c|c", "out")?.graph,
-            vec![("x", "x")],
-        );
-        let expected_table = graph.truth_table()?;
-        let mut transform = LogicGraphTransformer::new(graph);
-
-        transform.fold_or_chains()?;
-
-        let or_nodes = transform
-            .graph
-            .nodes
-            .iter()
-            .filter(|node| {
-                matches!(&node.kind, GraphNodeKind::Logic(logic) if logic.logic_type == LogicType::Or)
-            })
-            .collect_vec();
-        assert_eq!(or_nodes.len(), 2);
-        assert!(or_nodes.iter().all(|node| node.inputs.len() <= 2));
-        assert!(or_nodes.iter().all(|node| {
-            let mut inputs = node.inputs.clone();
-            inputs.sort_unstable();
-            inputs.dedup();
-            inputs.len() == node.inputs.len()
-        }));
-        assert_eq!(transform.graph.truth_table()?, expected_table);
-        transform.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn compose_high_level_gates_recovers_and_from_prepared_graph() -> eyre::Result<()> {
-        let graph = LogicGraph::from_stmt("a&b", "out")?.prepare_place()?;
-        let expected_table = graph.truth_table()?;
-        let mut transform = LogicGraphTransformer::new(graph);
-
-        transform.compose_high_level_gates()?;
-
-        assert_eq!(count_logic_type(&transform.graph, LogicType::And), 1);
-        assert_eq!(transform.graph.truth_table()?, expected_table);
-        transform.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn compose_high_level_gates_recovers_xor_from_prepared_graph() -> eyre::Result<()> {
-        let graph = LogicGraph::from_stmt("a^b", "out")?.prepare_place()?;
-        let expected_table = graph.truth_table()?;
-        let mut transform = LogicGraphTransformer::new(graph);
-
-        transform.compose_high_level_gates()?;
-
-        assert_eq!(count_logic_type(&transform.graph, LogicType::Xor), 1);
-        assert_eq!(transform.graph.truth_table()?, expected_table);
-        transform.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn compose_high_level_gates_handles_world_logic_graph() -> eyre::Result<()> {
-        let nbt = NBTRoot::load("test/alu.nbt")?;
-        let raw_world_graph = WorldGraphBuilder::new(&nbt.to_world()).build();
-        let transformer = WorldToLogicTransformer::new(raw_world_graph, true)?;
-        let logic_graph = transformer.transform()?;
-        let mut transform = LogicGraphTransformer::new(logic_graph);
-
-        transform.remove_double_neg_expression();
-        transform.optimize_cse()?;
-        transform.fold_or_chains()?;
-        transform.optimize_cse()?;
-        transform.compose_high_level_gates()?;
-        transform.optimize_cse()?;
-
-        assert!(count_logic_type(&transform.graph, LogicType::And) > 0);
-        transform.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn compose_high_level_gates_recovers_xor_from_world_xor_nbt() -> eyre::Result<()> {
-        let nbt = NBTRoot::load("test/xor-generated.nbt")?;
-        let raw_world_graph = WorldGraphBuilder::new(&nbt.to_world()).build();
-        let transformer = WorldToLogicTransformer::new(raw_world_graph, true)?;
-        let logic_graph = transformer.transform()?;
-        let mut transform = LogicGraphTransformer::new(logic_graph);
-
-        transform.remove_double_neg_expression();
-        transform.optimize_cse()?;
-        transform.fold_or_chains()?;
-        transform.optimize_cse()?;
-        transform.compose_high_level_gates()?;
-        transform.optimize_cse()?;
-
-        assert_eq!(count_logic_type(&transform.graph, LogicType::Xor), 1);
-        transform.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn compose_high_level_gates_recovers_xor_from_world_half_adder_nbt() -> eyre::Result<()> {
-        let nbt = NBTRoot::load("test/half-adder-generated.nbt")?;
-        let raw_world_graph = WorldGraphBuilder::new(&nbt.to_world()).build();
-        let transformer = WorldToLogicTransformer::new(raw_world_graph, true)?;
-        let logic_graph = transformer.transform()?;
-        let mut transform = LogicGraphTransformer::new(logic_graph);
-
-        transform.remove_double_neg_expression();
-        transform.optimize_cse()?;
-        transform.fold_or_chains()?;
-        transform.optimize_cse()?;
-        transform.compose_high_level_gates()?;
-        transform.optimize_cse()?;
-
-        assert!(count_logic_type(&transform.graph, LogicType::Xor) > 0);
-        assert_logic_nodes_have_input_count(&transform.graph, LogicType::Xor, 2);
-        transform.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn compose_high_level_gates_recovers_xor_from_world_full_adder_nbt() -> eyre::Result<()> {
-        let nbt = NBTRoot::load("test/full-adder.nbt")?;
-        let raw_world_graph = WorldGraphBuilder::new(&nbt.to_world()).build();
-        let transformer = WorldToLogicTransformer::new(raw_world_graph, true)?;
-        let logic_graph = transformer.transform()?;
-        let mut transform = LogicGraphTransformer::new(logic_graph);
-
-        transform.remove_double_neg_expression();
-        transform.optimize_cse()?;
-        transform.fold_or_chains()?;
-        transform.optimize_cse()?;
-        transform.compose_high_level_gates()?;
-        transform.optimize_cse()?;
-
-        assert!(count_logic_type(&transform.graph, LogicType::Xor) > 0);
-        assert_logic_nodes_have_input_count(&transform.graph, LogicType::Xor, 2);
-        transform.graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn prepare_place_runs_cse_for_buffered_full_adder() -> eyre::Result<()> {
-        let graph = predefined_logics::buffered_full_adder_graph()?;
-
-        assert!(!has_duplicate_expression(&graph, "Not(Input(a))"));
-        assert!(!has_duplicate_expression(&graph, "Not(Input(b))"));
-        assert!(!has_duplicate_expression(&graph, "Not(Input(cin))"));
-        graph.graph.verify()?;
-
-        Ok(())
-    }
-
-    fn count_logic_type(graph: &LogicGraph, logic_type: LogicType) -> usize {
-        graph
-            .nodes
-            .iter()
-            .filter(|node| {
-                matches!(&node.kind, GraphNodeKind::Logic(logic) if logic.logic_type == logic_type)
-            })
-            .count()
-    }
-
-    fn assert_logic_nodes_have_input_count(
-        graph: &LogicGraph,
-        logic_type: LogicType,
-        expected_input_count: usize,
-    ) {
-        for node in graph.nodes.iter().filter(|node| {
-            matches!(&node.kind, GraphNodeKind::Logic(logic) if logic.logic_type == logic_type)
-        }) {
-            assert_eq!(
-                node.inputs.len(),
-                expected_input_count,
-                "{logic_type:?} node {} has wrong input count",
-                node.id
-            );
-        }
-    }
-
-    #[test]
-    fn prepare_place_preserves_sequential_boundaries() -> eyre::Result<()> {
-        let mut graph = Graph::from_nodes(vec![
-            GraphNode {
-                kind: GraphNodeKind::Input("s".to_owned()),
-                outputs: vec![2],
-                ..Default::default()
-            },
-            GraphNode {
-                kind: GraphNodeKind::Input("r".to_owned()),
-                outputs: vec![2],
-                ..Default::default()
-            },
-            GraphNode {
-                kind: GraphNodeKind::Sequential(SequentialPrimitive::new(
-                    SequentialType::RsLatch,
-                    vec!["s".to_owned(), "r".to_owned()],
-                    vec!["q".to_owned()],
-                )),
-                inputs: vec![0, 1],
-                outputs: vec![3],
-                ..Default::default()
-            },
-            GraphNode {
-                kind: GraphNodeKind::Logic(Logic {
-                    logic_type: LogicType::Not,
-                }),
-                inputs: vec![2],
-                outputs: vec![4],
-                ..Default::default()
-            },
-            GraphNode {
-                kind: GraphNodeKind::Output("not_q".to_owned()),
-                inputs: vec![3],
-                ..Default::default()
-            },
-        ]);
-        graph.build_inputs();
-        graph.build_outputs();
-
-        let graph = LogicGraph { graph }.prepare_place()?;
-
-        assert_eq!(
-            graph
-                .nodes
-                .iter()
-                .filter(|node| node.kind.is_sequential())
-                .count(),
-            1
-        );
-        assert!(graph.nodes.iter().any(|node| matches!(
-            &node.kind,
-            GraphNodeKind::Logic(logic) if logic.logic_type == LogicType::Not
-        )));
-        graph.graph.verify()?;
-
-        Ok(())
     }
 }
